# backend/docs/swagger.yaml
basePath: /
definitions:
...
  piemdm_internal_model.NotificationLog:
    properties:
      ApprovalID:
        description: 审批实例ID
        type: string
      Content:
        description: 通知内容
        type: string
      CreatedAt:
        type: string
      ErrorMessage:
        description: 错误信息
        type: string
      ExtraData:
        description: 额外数据JSON
        type: string
      ID:
        type: integer
      MaxRetryCount:
        description: 最大重试次数
        type: integer
      NextRetryTime:
        description: 下次重试时间
        type: string
      NotificationType:
        description: 通知类型
        maxLength: 16
        type: string
      RecipientID:
        description: 接收人ID
        maxLength: 64
        type: string
      RecipientType:
        description: 接收人类型
        maxLength: 16
        type: string
      RetryCount:
        description: 重试次数
        type: integer
      SendTime:
        description: 发送时间
        type: string
      Status:
        description: 发送状态
        type: string
      TaskID:
        description: 任务ID
        type: string
      TemplateCode:
        description: 模板编码
        type: string
      TemplateID:
        description: 模板ID
        type: string
      Title:
        description: 通知标题
        maxLength: 200
        type: string
      UpdatedAt:
        type: string
    required:
    - Content
    - NotificationType
    - RecipientID
    - RecipientType
    - Title
    type: object
...
  piemdm_internal_service.ApprovalNotificationRequest:
    properties:
      approval_id:
        description: 审批实例ID
        type: string
      expire_time:
        description: 过期时间
        type: string
      notification_type:
        description: 通知类型
        type: string
      priority:
        description: 优先级
        type: integer
      recipient_id:
        description: 接收人ID
        type: string
      recipient_type:
        description: 接收人类型
        type: string
      task_id:
        description: 任务ID（可选）
        type: string
      template_type:
        description: 模板类型
        type: string
      variables:
        additionalProperties: {}
        description: 模板变量
        type: object
    type: object
...
  piemdm_internal_service.TemplateNotificationRequest:
    properties:
      approval_id:
        description: 审批实例ID（可选）
        type: string
      expire_time:
        description: 过期时间
        type: string
      notification_type:
        description: 通知类型
        type: string
      priority:
        description: 优先级
        type: integer
      recipient_id:
        description: 接收人ID
        type: string
      recipient_type:
        description: 接收人类型
        type: string
      task_id:
        description: 任务ID（可选）
        type: string
      template_code:
        description: 模板编码
        type: string
      variables:
        additionalProperties: {}
        description: 模板变量
        type: object
    type: object
  piemdm_internal_service.TestNotificationRequest:
    properties:
      content:
        description: 内容
        type: string
      notification_type:
        description: 通知类型
        type: string
      recipient_id:
        description: 接收人ID
        type: string
      title:
        description: 标题
        type: string
      variables:
        additionalProperties: {}
        description: 变量
        type: object
    type: object
...
host: localhost:8787
info:
  contact:
    email: support@piemdm.com
    name: API Support
  description: PieMDM Master Data Management System API Documentation
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: PieMDM API
  version: "1.0"
paths:
...
  /api/v1/notification_logs:
    get:
      consumes:
      - application/json
      description: 分页获取通知日志列表
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 15
        description: 每页数量
        in: query
        name: pageSize
        type: integer
      - description: 审批ID
        in: query
        name: approvalId
        type: string
      - description: 任务ID
        in: query
        name: taskId
        type: string
      - description: 接收人ID
        in: query
        name: recipientId
        type: string
      - description: 通知类型
        in: query
        name: notificationType
        type: string
      - description: 发送状态
        in: query
        name: status
        type: string
      - description: 开始时间
        in: query
        name: startTime
        type: string
      - description: 结束时间
        in: query
        name: endTime
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 获取通知日志列表
      tags:
      - 通知日志管理
  /api/v1/notification_logs/{id}:
    get:
      consumes:
      - application/json
      description: 根据ID获取通知日志详情
      parameters:
      - description: 日志ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/piemdm_internal_model.NotificationLog'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 获取通知日志详情
      tags:
      - 通知日志管理
...
  /api/v1/notifications/approval:
    post:
      consumes:
      - application/json
      description: 发送审批相关的通知
      parameters:
      - description: 通知信息
        in: body
        name: notification
        required: true
        schema:
          $ref: '#/definitions/piemdm_internal_service.ApprovalNotificationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 发送审批通知
      tags:
      - 通知管理
  /api/v1/notifications/batch:
    post:
      consumes:
      - application/json
      description: 批量发送多个通知
      parameters:
      - description: 通知列表
        in: body
        name: notifications
        required: true
        schema:
          items:
            $ref: '#/definitions/piemdm_internal_service.ApprovalNotificationRequest'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 批量发送通知
      tags:
      - 通知管理
  /api/v1/notifications/process-pending:
    post:
      consumes:
      - application/json
      description: 手动触发处理待发送的通知
      parameters:
      - default: 100
        description: 处理数量限制
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 处理待发送通知
      tags:
      - 通知管理
  /api/v1/notifications/retry-failed:
    post:
      consumes:
      - application/json
      description: 处理重试发送失败的通知
      parameters:
      - default: 50
        description: 重试数量限制
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 处理重试通知
      tags:
      - 通知管理
  /api/v1/notifications/statistics:
    get:
      consumes:
      - application/json
      description: 获取通知发送统计信息
      parameters:
      - description: 开始时间
        in: query
        name: startTime
        type: string
      - description: 结束时间
        in: query
        name: endTime
        type: string
      - description: 分组方式
        enum:
        - day
        - week
        - month
        in: query
        name: groupBy
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 获取通知统计
      tags:
      - 通知管理
  /api/v1/notifications/template:
    post:
      consumes:
      - application/json
      description: 使用指定模板发送通知
      parameters:
      - description: 模板通知信息
        in: body
        name: notification
        required: true
        schema:
          $ref: '#/definitions/piemdm_internal_service.TemplateNotificationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 使用模板发送通知
      tags:
      - 通知管理
  /api/v1/notifications/test:
    post:
      consumes:
      - application/json
      description: 测试通知渠道是否正常工作
      parameters:
      - description: 测试信息
        in: body
        name: test
        required: true
        schema:
          $ref: '#/definitions/piemdm_internal_service.TestNotificationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 测试通知
      tags:
      - 通知管理
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"